+++ export_test.go
@@ -5,6 +5,7 @@ import (
 	"math/rand"
 	"testing"
 
+	"cosmossdk.io/log"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
@@ -14,7 +15,7 @@ import (
 // setupExportTreeBasic sets up a basic tree with a handful of
 // create/update/delete operations over a few versions.
 func setupExportTreeBasic(t require.TestingT) *ImmutableTree {
-	tree, err := NewMutableTree(db.NewMemDB(), 0, false)
+	tree, err := NewMutableTree(db.NewMemDB(), 0, false, log.NewNopLogger())
 	require.NoError(t, err)
 
 	_, err = tree.Set([]byte("x"), []byte{255})
@@ -74,7 +75,7 @@ func setupExportTreeRandom(t *testing.T) *ImmutableTree {
 	)
 
 	r := rand.New(rand.NewSource(randSeed))
-	tree, err := NewMutableTree(db.NewMemDB(), 0, false)
+	tree, err := NewMutableTree(db.NewMemDB(), 0, false, log.NewNopLogger())
 	require.NoError(t, err)
 
 	var version int64
@@ -134,7 +135,7 @@ func setupExportTreeSized(t require.TestingT, treeSize int) *ImmutableTree { //n
 	)
 
 	r := rand.New(rand.NewSource(randSeed))
-	tree, err := NewMutableTree(db.NewMemDB(), 0, false)
+	tree, err := NewMutableTree(db.NewMemDB(), 0, false, log.NewNopLogger())
 	require.NoError(t, err)
 
 	for i := 0; i < treeSize; i++ {
@@ -229,7 +230,7 @@ func TestExporterCompress(t *testing.T) {
 
 func TestExporter_Import(t *testing.T) {
 	testcases := map[string]*ImmutableTree{
-		"empty tree": NewImmutableTree(db.NewMemDB(), 0, false),
+		"empty tree": NewImmutableTree(db.NewMemDB(), 0, false, log.NewNopLogger()),
 		"basic tree": setupExportTreeBasic(t),
 	}
 	if !testing.Short() {
@@ -256,7 +257,7 @@ func TestExporter_Import(t *testing.T) {
 					exporter = NewCompressExporter(innerExporter)
 				}
 
-				newTree, err := NewMutableTree(db.NewMemDB(), 0, false)
+				newTree, err := NewMutableTree(db.NewMemDB(), 0, false, log.NewNopLogger())
 				require.NoError(t, err)
 				innerImporter, err := newTree.Import(tree.Version())
 				require.NoError(t, err)
@@ -327,7 +328,7 @@ func TestExporter_Close(t *testing.T) {
 }
 
 func TestExporter_DeleteVersionErrors(t *testing.T) {
-	tree, err := NewMutableTree(db.NewMemDB(), 0, false)
+	tree, err := NewMutableTree(db.NewMemDB(), 0, false, log.NewNopLogger())
 	require.NoError(t, err)
 
 	_, err = tree.Set([]byte("a"), []byte{1})