+++ nodedb.go
@@ -11,12 +11,12 @@ import (
 	"strings"
 	"sync"
 
+	"cosmossdk.io/log"
 	dbm "github.com/cosmos/cosmos-db"
 
 	"github.com/cosmos/iavl/cache"
 	"github.com/cosmos/iavl/fastnode"
 	ibytes "github.com/cosmos/iavl/internal/bytes"
-	"github.com/cosmos/iavl/internal/logger"
 	"github.com/cosmos/iavl/keyformat"
 )
 
@@ -60,6 +60,8 @@ var (
 var errInvalidFastStorageVersion = fmt.Sprintf("Fast storage version must be in the format <storage version>%s<latest fast cache version>", fastStorageVersionDelimiter)
 
 type nodeDB struct {
+	logger log.Logger
+
 	mtx            sync.Mutex       // Read/write lock.
 	db             dbm.DB           // Persistent node storage.
 	batch          dbm.Batch        // Batched writing buffer.
@@ -72,7 +74,7 @@ type nodeDB struct {
 	fastNodeCache  cache.Cache      // Cache for nodes in the fast index that represents only key-value pairs at the latest version.
 }
 
-func newNodeDB(db dbm.DB, cacheSize int, opts *Options) *nodeDB {
+func newNodeDB(db dbm.DB, cacheSize int, opts *Options, lg log.Logger) *nodeDB {
 	if opts == nil {
 		o := DefaultOptions()
 		opts = &o
@@ -85,6 +87,7 @@ func newNodeDB(db dbm.DB, cacheSize int, opts *Options) *nodeDB {
 	}
 
 	return &nodeDB{
+		logger:         lg,
 		db:             db,
 		batch:          db.NewBatch(),
 		opts:           *opts,
@@ -198,7 +201,7 @@ func (ndb *nodeDB) SaveNode(node *Node) error {
 		}
 	}
 
-	logger.Debug("BATCH SAVE %+v\n", node)
+	ndb.logger.Debug("BATCH SAVE", "node", node)
 	ndb.nodeCache.Add(node)
 	return nil
 }
@@ -388,10 +391,7 @@ func (ndb *nodeDB) DeleteVersionsFrom(fromVersion int64) error {
 
 	// Delete the nodes
 	err = ndb.traverseRange(nodeKeyFormat.Key(fromVersion), nodeKeyFormat.Key(latest+1), func(k, v []byte) error {
-		if err = ndb.batch.Delete(k); err != nil {
-			return err
-		}
-		return nil
+		return ndb.batch.Delete(k)
 	})
 
 	if err != nil {
@@ -579,11 +579,7 @@ func (ndb *nodeDB) traverseRange(start []byte, end []byte, fn func(k, v []byte) 
 		}
 	}
 
-	if err := itr.Error(); err != nil {
-		return err
-	}
-
-	return nil
+	return itr.Error()
 }
 
 // Traverse all keys with a certain prefix. Return error if any, nil otherwise
@@ -776,33 +772,38 @@ func (ndb *nodeDB) size() int {
 	return size
 }
 
-func (ndb *nodeDB) traverseNodes(fn func(node *Node) error) error {
-	ndb.resetLatestVersion(0)
-	latest, err := ndb.getLatestVersion()
-	if err != nil {
-		return err
+func isReferenceToRoot(bz []byte) bool {
+	if bz[0] == nodeKeyFormat.Prefix()[0] {
+		if len(bz) == 13 {
+			return true
+		}
 	}
+	return false
+}
 
+func (ndb *nodeDB) traverseNodes(fn func(node *Node) error) error {
 	nodes := []*Node{}
-	for version := int64(1); version <= latest; version++ {
-		if err := ndb.traverseRange(nodeKeyFormat.Key(version), nodeKeyFormat.Key(version+1), func(key, value []byte) error {
-			var (
-				version int64
-				nonce   int32
-			)
-			nodeKeyFormat.Scan(key, &version, &nonce)
-			node, err := MakeNode(&NodeKey{
-				version: version,
-				nonce:   nonce,
-			}, value)
-			if err != nil {
-				return err
-			}
-			nodes = append(nodes, node)
+
+	if err := ndb.traversePrefix(nodeKeyFormat.Key(), func(key, value []byte) error {
+		if isReferenceToRoot(value) {
 			return nil
-		}); err != nil {
+		}
+		var (
+			version int64
+			nonce   int32
+		)
+		nodeKeyFormat.Scan(key, &version, &nonce)
+		node, err := MakeNode(&NodeKey{
+			version: version,
+			nonce:   nonce,
+		}, value)
+		if err != nil {
 			return err
 		}
+		nodes = append(nodes, node)
+		return nil
+	}); err != nil {
+		return err
 	}
 
 	sort.Slice(nodes, func(i, j int) bool {
@@ -849,7 +850,7 @@ func (ndb *nodeDB) traverseStateChanges(startVersion, endVersion int64, fn func(
 
 		var changeSet ChangeSet
 		receiveKVPair := func(pair *KVPair) error {
-			changeSet.Pairs = append(changeSet.Pairs, *pair)
+			changeSet.Pairs = append(changeSet.Pairs, pair)
 			return nil
 		}